Revert "merge demov2 into giangnguyen"
This reverts commit c79c9fcbcac32ac07b3a3636e2b10cc52a925952, reversing
changes made to b83373a80d505897762b16143368372eff587ed0.

+++ src/main/java/fptu/fcharity/entity/Wallet.java
@@ -6,7 +6,6 @@ import lombok.Setter;
 import org.hibernate.annotations.ColumnDefault;
 import org.hibernate.annotations.Nationalized;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 @Getter
@@ -22,6 +21,6 @@ public class Wallet {
 
     @Nationalized
     @Column(name = "balance")
-    private BigDecimal balance;
+    private int balance;
 
 }
